requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
python-dotenv
docx2txt
PyMuPDF
tiktoken